requirements.txt
annotated-types==0.6.0
boto3==1.34.68
botocore==1.34.68
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
jmespath==1.0.1
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
typing_extensions==4.9.0
urllib3==2.2.1
